Refactor TableTest assertion methods to a separate utility class
